# api/internal/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: ZaaS - Zufall as a Service
  description: |
    **Zufall as a Service** - free randomness API. Roll dice, generate UUIDs, pick
    random colors, get random geographic coordinates, generate passwords, random words,
    lorem ipsum, SSH keypairs, and more.

    **Base URL:** `https://zaas.at/api/v1`

    **Authentication:** All endpoints are usable anonymously. Provide an
    `Authorization: Bearer <api_key>` header to use your registered API key and
    receive a higher rate limit. Obtain an API key via `POST /auth/register`.

    **Rate limiting:** Anonymous requests are rate-limited per IP. Authenticated
    requests are rate-limited per API key. The current limits are returned in
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers.
    A `429` response includes a `Retry-After` header.

    **Response envelope:** Every successful response is wrapped in
    `{result, meta}` (single) or `{results, meta}` (multiple). The `meta` object
    echoes the resolved request parameters, the request ID, and a timestamp for
    reproducibility and support triage.
  version: "1.0.0"
  license:
    name: MIT
  termsOfService: https://zaas.at/imprint
  contact:
    name: ZaaS
    url: https://zaas.at
    email: hello@zaas.at
servers:
  - url: https://zaas.at/api/v1
    description: Production
security:
  - {}
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [Auth]
      summary: Register for an API key
      description: Submit email and display name to receive a verification email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "202":
          description: Accepted - verification email sent (always returns 202 to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          $ref: "#/components/responses/InvalidParam"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/verify:
    post:
      operationId: authVerify
      tags: [Auth]
      summary: Verify email and receive API key
      description: Consumes the one-time verification token from the email link and returns an API key. Must be called via a POST (e.g. from the /verify landing page) to prevent mail scanners from consuming the token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verification successful - API key returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          $ref: "#/components/responses/InvalidParam"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/reissue:
    post:
      operationId: authReissue
      tags: [Auth]
      summary: Request API key re-issue
      description: Submit email to receive a re-issue verification email. Old key will be revoked upon verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReissueRequest"
      responses:
        "202":
          description: Accepted - re-issue verification email sent (always returns 202)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          $ref: "#/components/responses/InvalidParam"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, display_name]
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
          minLength: 1
          maxLength: 100
    VerifyRequest:
      type: object
      required: [token, type]
      properties:
        token:
          type: string
          description: One-time verification token from the email link
        type:
          type: string
          enum: [registration, reissue]
          description: Token type matching the registration or reissue flow
    ReissueRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AcceptedResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    VerifyResponse:
      type: object
      required: [api_key, message]
      properties:
        api_key:
          type: string
        message:
          type: string
    Problem:
      type: object
      description: "RFC 9457 Problem Details (https://datatracker.ietf.org/doc/html/rfc9457)"
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://zaas.at/errors/invalid-param"
        title:
          type: string
          description: Short human-readable summary of the problem type
          example: "Bad Request"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation of this specific occurrence
          example: "sides must be one of 4,6,8,10,12,20,100"
        instance:
          type: string
          description: URI reference identifying the specific request that caused the problem
          example: "/api/v1/dice"
        code:
          type: string
          enum: [RATE_LIMITED, INVALID_PARAM, INTERNAL_ERROR, INVALID_API_KEY, SERVICE_UNAVAILABLE]
          description: Machine-readable error code (ZaaS extension)
        retry_after:
          type: integer
          description: Seconds to wait before retrying (ZaaS extension, only present for RATE_LIMITED)
        request_id:
          type: string
          description: Unique request identifier for support triage (ZaaS extension)
  responses:
    InvalidParam:
      description: Invalid query parameter
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServiceUnavailable:
      description: Service temporarily unavailable (e.g. database not configured)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key in format: zaas_<32 hex chars>"